# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒体同步和刮削系统API",
        "title": "QMediaSync API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8115",
    "basePath": "/",
    "paths": {
...
        "/update/cancel": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "取消正在进行的更新任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "更新管理"
                ],
                "summary": "取消更新",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/update/last": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取GitHub上最新的5个稳定版本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "更新管理"
                ],
                "summary": "获取最新版本",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/update/progress": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询当前更新任务的下载和安装进度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "更新管理"
                ],
                "summary": "获取更新进度",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}